refactor(discord/events): relocate internal client + schemas + URLs per domain
