+++ backend/discordbot/schemas.py
@@ -1,7 +1,16 @@
-"""Pydantic schemas for Discord bot API data transfer.
+"""Pydantic REQUEST schemas for the Discord bot's internal-API surface.
 
-These schemas validate data returned by internal API endpoints
-consumed by the Discord bot (via celery tasks or direct HTTP calls).
+This file owns:
+  * Schemas validating data returned by internal API endpoints the bot reads
+    (``SyncDiscordStateSchema``, ``MessageLogSchema``, ``DiscordEventStateSchema``).
+  * Request-body validators for the bot's signup-flow endpoints
+    (``discordbot/internal_signup_views.py``) — they validate Discord-button
+    transport payloads so they belong here with the bot.
+
+RESPONSE schemas for those signup endpoints (``SignupActionResponse``,
+``RankFlowStateResponse``) live in ``events/schemas.py`` since they are the
+contract of the ``events.discord.handlers`` business logic, not the bot's
+transport layer.
 """
 
 from __future__ import annotations
@@ -149,71 +158,4 @@ class SetPositionRequest(BaseModel):
 class RankFlowStateRequest(BaseModel):
     event_id: int
     discord_user_id: str
-    model_config = {"extra": "ignore"}
-
-
-# ---------------------------------------------------------------------------
-# Internal signup-API response schema (validated client-side in
-# app/internal_client/signup_actions.py).
-# ---------------------------------------------------------------------------
-
-
-class SignupActionResponse(BaseModel):
-    """All bot-facing handler return dicts share this shape.
-
-    Different handlers populate different subsets; consumer code branches on
-    ``action`` and reads only the relevant fields. Used as the union envelope
-    so a single pydantic model can validate every signup-flow response.
-    """
-
-    # Common fields
-    action: Optional[str] = None  # signed_up | needs_modal | needs_screenshot |
-    # needs_rank_details | needs_rank_status | error | tentative | declined |
-    # not_signed_up | already_declined | already_tentative | positions_saved |
-    # position_set
-    status: Optional[str] = None  # SignupStatus value when action == "signed_up"
-    message: Optional[str] = None  # human-readable for error / not_signed_up / etc.
-
-    # signup_modal_submit / signup_button: needs_modal payload
-    game_type: Optional[int] = None
-    prefill: Optional[dict] = None
-    require_steam_id: Optional[bool] = None
-    require_rank_screenshot: Optional[bool] = None
-    require_battlecup_screenshot: Optional[bool] = None
-    min_mmr: Optional[int] = None
-    allow_active_mmr: Optional[bool] = None
-    allow_previous_rank: Optional[bool] = None
-    allow_battlecup_rating: Optional[bool] = None
-
-    # rank_medal_select / battle_cup_submit: needs_screenshot payload
-    screenshot_type: Optional[str] = None
-    medal: Optional[str] = None
-    tier: Optional[str] = None
-
-    # notify_button
-    subscribed: Optional[bool] = None
-
-    # screenshot_upload
-    success: Optional[bool] = None
-    signed_up: Optional[bool] = None
-
-    # save_positions
-    positions: Optional[list[int]] = None
-
-    model_config = {"extra": "ignore"}
-
-
-class RankFlowStateResponse(BaseModel):
-    """Response of /api/internal/discord/rank-flow-state/.
-
-    Used by the legacy pos_confirm flow in bot.py to construct the next
-    RankDetailsView. ``error``/``message`` populated only on lookup failure;
-    otherwise the rank_status + require_screenshot + min_mmr fields drive UI.
-    """
-
-    rank_status: Optional[str] = None  # "active" | "previous" | "never"
-    require_screenshot: bool = False
-    min_mmr: Optional[int] = None
-    error: Optional[str] = None  # set when the lookup fails (e.g. event_not_found / no_org_user)
-    message: Optional[str] = None
     model_config = {"extra": "ignore"}
